Don't attempt to delete file if it does not exist

+++ collectfast/management/commands/collectstatic.py
@@ -119,7 +119,8 @@ class Command(collectstatic.Command):
                 path, prefixed_path, source_storage)
         if not self.dry_run:
             self.log("Deleting '%s'" % path)
-            self.storage.delete(prefixed_path)
+            if self.storage.exists(prefixed_path):
+                self.storage.delete(prefixed_path)
         else:
             self.log("Pretending to delete '%s'" % path)
         return True
